ChatPdf/requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
langchain-community
google-generativeai
python-dotenv
faiss-cpu
